pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "graphql-agent"
version = "0.1.0"
description = "GraphQL prompt routing server built with Strawberry"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "YHardik Ojha" }]
dependencies = [
  "strawberry-graphql>=0.233",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "starlette>=0.50.0",
  "openai>=1.0",
  "google-genai>=1.49.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.14.4",
  "mypy>=1.11"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
# uv uses standard PEP 621 metadata; this section is optional and can be extended later

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
ignore = [
  # Allow module-level variable annotations without assignment in configs
  "ANN101",
  "E501",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]


[dependency-groups]
dev = [
    "mypy>=1.18.2",
]


